feat: P4ADEV-4559 send timeline consumer, changed recipientIndex from Long to Integer.

+++ src/main/java/it/gov/pagopa/pu/registry/dto/RegistryEventSendTimelineDTO.java
@@ -25,7 +25,7 @@ public class RegistryEventSendTimelineDTO extends RegistryInterfaceEventDTO {
   @NotNull
   private String notificationRequestId;
   private String iun;
-  private Long recipientIndex;
+  private Integer recipientIndex;
   private String newStatus;
   @NotNull
   private RegistryOutcome outcome;
